# docs/openapi.yaml
openapi: 3.1.0
info:
  title: NexusTale API
  version: "1.0.0"
  description: >
    REST API for NexusTale — a sci-fi/fantasy novel-writing platform with
    Git-backed versioning, world wiki, and AI co-author support.
servers:
  - url: /api/v1
    description: Current API prefix (served by nginx reverse-proxy in all environments)
security: []  # default unauthenticated; individual operations override via bearerAuth
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Reusable error shapes ───────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid JWT.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    BadRequest:
      description: Validation or binding error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }

    # ── Auth ──────────────────────────────────────────────────────────────────
...
    EntityType:
      type: string
      enum: [character, location, faction, item, concept, lore]
    CreateEntityRequest:
      type: object
      required: [type, name]
      properties:
        type:             { $ref: '#/components/schemas/EntityType' }
        name:             { type: string, minLength: 1, maxLength: 200 }
        summary:          { type: string }
        attributes:       { type: object, additionalProperties: { type: string } }
        parent_entity_id: { type: string, format: uuid }
...
paths:
...
  /projects/{projectId}/wiki/entities:
    parameters:
      - name: projectId
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: type
        in: query
        schema: { $ref: '#/components/schemas/EntityType' }
        description: Filter by entity type.
    get:
      operationId: listEntities
      summary: List wiki entities, optionally filtered by type
      tags: [wiki]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: List of entities.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EntityResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createEntity
      summary: Create a new wiki entity
      tags: [wiki]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateEntityRequest' }
      responses:
        "201":
          description: Entity created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EntityResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /projects/{projectId}/wiki/entities/{entityId}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: entityId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getEntity
      summary: Get a single wiki entity
      tags: [wiki]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Entity details.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EntityResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
    patch:
      operationId: updateEntity
      summary: Update a wiki entity
      tags: [wiki]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateEntityRequest' }
      responses:
        "200":
          description: Updated entity.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EntityResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteEntity
      summary: Delete a wiki entity
      tags: [wiki]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Entity deleted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
  /projects/{projectId}/wiki/entities/{entityId}/children:
    parameters:
      - name: projectId
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: entityId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: listChildEntities
      summary: List child entities of a parent entity
      tags: [wiki]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Child entities.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/EntityResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
    post:
      operationId: createChildEntity
      summary: Create a child entity under a parent
      tags: [wiki]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateEntityRequest' }
      responses:
        "201":
          description: Child entity created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EntityResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
  /projects/{projectId}/wiki/entities/{entityId}/image:
    parameters:
      - name: projectId
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: entityId
        in: path
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: uploadEntityImage
      summary: Upload or replace an entity portrait image
      tags: [wiki]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
      responses:
        "200":
          description: Entity with updated image_url.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EntityResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteEntityImage
      summary: Remove an entity's portrait image
      tags: [wiki]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Entity with image_url cleared.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/EntityResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }

  # ── Wiki — Relationships ──────────────────────────────────────────────────────
  /projects/{projectId}/wiki/relationships:
    parameters:
      - name: projectId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: listRelationships
      summary: List all entity relationships in the project
      tags: [wiki]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: List of relationships.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/RelationshipResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createRelationship
      summary: Create a relationship between two entities
      tags: [wiki]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateRelationshipRequest' }
      responses:
        "201":
          description: Relationship created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/RelationshipResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /projects/{projectId}/wiki/relationships/{relationshipId}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: relationshipId
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      operationId: deleteRelationship
      summary: Delete a relationship
      tags: [wiki]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Relationship deleted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }

  # ── Wiki — Graph ──────────────────────────────────────────────────────────────
  /projects/{projectId}/wiki/graph:
    parameters:
      - name: projectId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: getWikiGraph
      summary: Get all entities and relationships in one call (for graph rendering)
      tags: [wiki]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Full wiki graph.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WikiGraphResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }

  # ── Wiki — Magic Rules ────────────────────────────────────────────────────────
  /projects/{projectId}/wiki/magic-rules:
    parameters:
      - name: projectId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: listMagicRules
      summary: List all magic / world-system rules
      tags: [wiki]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: List of magic rules.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MagicRuleResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createMagicRule
      summary: Create a new magic rule
      tags: [wiki]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateMagicRuleRequest' }
      responses:
        "201":
          description: Magic rule created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MagicRuleResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /projects/{projectId}/wiki/magic-rules/{ruleId}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: ruleId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updateMagicRule
      summary: Update a magic rule
      tags: [wiki]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateMagicRuleRequest' }
      responses:
        "200":
          description: Updated magic rule.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MagicRuleResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteMagicRule
      summary: Delete a magic rule
      tags: [wiki]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Magic rule deleted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }

  # ── Wiki — Timeline Events ────────────────────────────────────────────────────
  /projects/{projectId}/wiki/timeline:
    parameters:
      - name: projectId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: listTimelineEvents
      summary: >
        List all timeline events with resolved dates. Anchored events
        have their year/month/day computed from their anchor chain.
      tags: [wiki]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Timeline events.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/TimelineEventResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
    post:
      operationId: createTimelineEvent
      summary: Create a new timeline event
      tags: [wiki]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTimelineEventRequest' }
      responses:
        "201":
          description: Timeline event created.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TimelineEventResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
  /projects/{projectId}/wiki/timeline/{eventId}:
    parameters:
      - name: projectId
        in: path
        required: true
        schema: { type: string, format: uuid }
      - name: eventId
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updateTimelineEvent
      summary: Update a timeline event
      tags: [wiki]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateTimelineEventRequest' }
      responses:
        "200":
          description: Updated timeline event.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TimelineEventResponse' }
        "400": { $ref: '#/components/responses/BadRequest' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }
    delete:
      operationId: deleteTimelineEvent
      summary: Delete a timeline event
      tags: [wiki]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Timeline event deleted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MessageResponse' }
        "401": { $ref: '#/components/responses/Unauthorized' }
        "404": { $ref: '#/components/responses/NotFound' }

  # ── User account ─────────────────────────────────────────────────────────────
...
  /projects/{projectId}/wiki/autolink:
    parameters:
      - name: projectId
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      operationId: autolink
      summary: Find entity name mentions in a block of text
      tags: [wiki]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: text
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: List of entity matches with character offsets.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/AutolinkMatch' }
        "401": { $ref: '#/components/responses/Unauthorized' }

  # ── Guide ────────────────────────────────────────────────────────────────────
...